[Bug 18893] Fix formatting in list of stack modes
The table of possible values for the `mode` property had lost its
formatting.  I experimented with displaying the values as a table, but
found that it was hard to give it a good appearance: the number and
summary columns took up too much horizontal space, leaving the
detailed description crammed into a tall, narrow cell.

Instead, this patch reformats the information as a bullet point list
with headed paragraphs, which seems to be quite readable.