[Bug 13055] docs: Improve syntax description for "try"
The LiveCode 8+ dictionary is able to show multiline syntax quite well
in the main documentation display.  Although it is only able to show
the first line of the syntax in the list of dictionary entries, this
change seems to make the dictionary entry for the `try` control
structure a lot clearer.