lc-compile: Explicitly initialise some stack variables.
Several `BindingRef` stack variables were not initialised to `NULL` at
declaration.  If `FindNameInScope()` was working correctly, then this
wouldn't be a problem: it would be assigned before any use.  However,
`FindNameInScope()` was sometimes returning non-zero _without_
assigning its out parameter.

During `lc-compile` execution, the random value on the stack often
happened to be a valid memory address.  This coincidence masked the
fact that `FindNameInScope()` was not working correctly, and caused
weird compiler crashes.