[Bug 15604] docs: Update links to PCRE pattern documentation
When PCRE2 was introduced, the location of the PCRE documentation on
the PCRE website was changed, breaking the links in the dictionary.

This patch updates the dictionary to link directly to the
documentation of PCRE regular expression patterns, rather than to the
toplevel list of PCRE documentation.  Most of the PCRE documentation
is about the PCRE APIs, which aren't relevant to almost all LiveCode
users.

(cherry picked from commit 3e67d728d1642d637a77f4e1a0c0abfe9bd519e0)