[Bug 15117] docs: lineOffset() can find multiline strings
In recent versions of LiveCode, the `lineOffset()` function has been
enhanced to be able to search for multiline substrings.  It is no
longer the case that it will return 0 when the `lineToFind` contains
line separators. This patch removes that statement.

(cherry picked from commit 98479573bfe251ab60893d8695f1e2719f9a2a0a)