[Bug 14172] docs: "combine" uses lexicographic key ordering
The `combine` command was previously documented as producing a result
based on internal hash order of the input array's keys.  However, the
ordering is in fact in case-sensitive lexicographic order.  This patch
corrects the incorrect note and provides an example of a script that
places the keys in an equivalent ordering.