[Bug 14473] docs: Provide a complete revZipAddItemWithData example
Argument 3 passed to the `revZipAddItemWithData` command is required
to be the name of a variable containing the data to be stored.  The
previous example 2 of the command was incorrect, attempting to pass a
string literal to be stored as the content as argument 3.

This patch removes the incorrect example, and makes the original,
correct example more complete by providing a full example of how to
use `revZipAddItemWithData` to create a UTF-8 encoded text file in the
zip archive.

(cherry picked from commit 85dd4d637ae337bee803a03411caad459431c36b)