[Bug 14247] docs: Correct info on field insertion point on focus
When the `focus` command is used to give keyboard focus to a field,
the insertion point in the field is restored to its location when the
field was last focused.  The `focus` command does not move the
insertion point to the end of the field's contents.

(cherry picked from commit 2bd801ee8ec5692fbd68f6336261a4c6d1c0f75b)