[Bug 18853] libscript: Add support for loading multiple modules at once
In commit 221897f, `lc-run` gained the ability to load multiple
modules from a single `.lcm` bytecode file.  This commit moves the
implementation from `lc-run` itself into libscript.

To make the new API easy to use, it makes use of managed lifetime
types in its prototype.  This requires it to be placed in
`script-auto.h`, because the required types aren't available in
`script.h`.