Add kMCEmptyRectangle constant to workaround missing C++11 features
There are several places in the source code where uninitialised
`MCRectangle` values cause Coverity to detect codepaths which make use
of unitialised values.

Ideally, `MCRectangle` would have a `constexpr` default constructor
that initialises its members to 0.

Unfortunately, that isn't feasible because it requires unrestricted
union support. For example, `MCRectangle` is used in a union in
`MCExecValue`.  Without C++11 unrestricted unions, types that are used
in a union must be POD and must not have constructors or destructors.

We will not be able to use constructors on `MCRectangle` and related
types until we update to MSVC 2017.  In the mean time, adding a
`kMCEmptyRectangle` constant allows RAII-like use of MCRectangle
values.