[[ Id Cache ]] Clear stack cache on delete before removing controls Object destructors use getstack() to fetch the stack from whose id cache they are removing themselves. Since getstack() returns the defaultStack if the parent is already deleted, this causes controls to attempt to remove ids from the cache of the default stack instead. This causes the default stack's id cache to become out of sync with the corresponding m_in_id_cache flag of its controls.
This patch combats the problem in two ways:
Firstly, move the destruction of a stack's ID cache to *above* the
removal of its controls in its destructor, and 'uncache' each control
properly when doing so. This ensures that when these controls'
destructors are called, their m_in_id_cache flags are already false.

This also makes the clearing of the object cache more efficient.

Secondly, ensure that when controls are removed from the id cache, it
is the object that was actually removed that has its flag set.