[Bug 19008] Improve modal placement
If neither the defaultStack or the topStack can provide valid
rectangles to place modal stack relative to the current behavior
will place modals relative to unopen defaultStacks. This patch adds
visiblity checks to the defaultStack and topStack and if neither
can provide a reasonable rectangle to place the modal relative to
it will be placed relative to its own rectangle. This change allows
an unopen defaultStack to set the location of a modal dialog
before presenting it.