revvideograbber: Remove from Mac OS X builds
revvideograbber for Mac OS X depends on QuickTime, which is not
available to build against when using the `macosx10.9` SDK (or newer).
In LiveCode 9, we will end support for Mac OS X 10.8 and earlier.  The
need to be able to build revvideograbber has been preventing from
using newer C++ language and standard library features, and has made
it difficult to get started with LiveCode engine development.

This patch discontinues revvideograbber for Mac, and updates all
revvideograbber documentation accordingly.  revvideograbber is still
supported on Windows, using DirectShow.  Some of its commands are no
longer useful, including `revVideoGrabSettings`,
`revSetVideoGrabSettings`, and `revVideoGrabIdle`.