[CID 16369] MCCard::relayer(): Restructure code to prevent possible memory leak
Coverity detected a possible codepath by which `MCCard::relayer()`
could leak a `MCObjptr` instance.  This patch restructures part of the
relayering code to first calculate an insertion position, then
allocate the `MCObjptr`, before inserting it.  It ensures that there
is no possible code path that permits a new instance to be allocated
without moving its ownership to the `MCCard` instance.

Coverity-ID: 16369