[[ Bug 19063 ]] Bind builtin foreign handlers to libscript's object
This patch changes the use of GetProcAddress() / dlsym() in libscript
so that it uses the shared object containing libscript, rather than
the main application shared object.

The handle for libscript's shared object is acquired in Initialize
and released in Finalize.

It can be queried at any time using the (libscript private) API
MCScriptGetModuleHandle.