[[ Bug 18977 ]] Don't short-cut card setting in no ui mode This was preventing openCard messages in no ui mode.
It also causes `MCCard::open` to be run when the only card on
a stack is deleted, in no ui mode. This causes any existing
controls in a card's `objptr`s to be reparented to the card,
causing a crash if a shared group is present, as this needs to
stay reparented to the stack.