[[ Foundation ]] Improve notion of index name
This patch improves the performance of names when used as
indexes into arrays. It adds an optimized codepath for looking
up and creating names which are integers (in the range supported
by index_t) and ensures these codepaths are used when converting
from a value to a nameref if the initial value is a number.