[CID 16934] ExternalV0::Handle(): Prevent memory leak
If an external handler argument could not be evaluated, or could not
be converted to a string, it was possible that the computed parameter
array could be leaked.

This patch fixes the leak by using a managed lifetime array, and
increases efficiency by allocating the array in a single shot rather
than calling `MCU_realloc()` to extend it for each parameter found.

It also fixes a possible null deference path.

Coverity-ID: 16934
Coverity-ID: 15189