[CID 17002] Use managed lifetime for MCObjectPropertySet::m_props
The internal `MCObjectPropertySet::m_props` array is now an
`MCAutoArrayRef`, which is created lazily the first time it is needed.
This allows `MCObjectPropertySet` to just use the default destructor.

A new private `fetch_nocopy()` member function has been added, which
allows `const` accessors to work as expected even if the `m_props`
field has not yet been populated.

This patch also fixes several issues where the contents of `m_prop`
could be lost if an error occurred (such as OOM).  This ensures that
operation failure leaves state unchanged.

Coverity-ID: 17002