[[ VS2015 ]] Tweak environment options for VS2015
This patch tweaks various environment flags and options to ensure
a build works under VS2015.

It ensures that the ARCH variable is always set and defaults to x86.
It maps ARCH to the Microsoft architecture name (X86 = x86 and X64 =
x86_64) and ensures that it is used to set the target_arch of the build.

To ensure that things build consistently targetting the correct
architecture, it ensures that /MACHINE is set for the link and
lib tools.

Finally it ensures the gyp is called targetting VS 2015 output and
the architecture specified by ARCH.