build: Replace config.sh script with Python script
In order to facilitate builds on Windows where a Bourne shell is
not easily available, this patch replaces the toplevel `config.sh`
script with an equivalent `config.py` Python script.  All the rest
of the build infrastructure is kept the same and all the same
commands will continue to work unchanged.