[CID 16332] Fix resource leak in MCServerScript::FindFileIndex()
`MCServerScript::FindFile()` looks for and returns the server script
file record for the specified filename, and optionally allocates a new
record if one was not found.

When a new record is allocated, it is _not_ added to the list of
records in the `MCServerScript` instance.  This requires the caller of
`FindFile()` to free or store the record.  However, there is no
_direct_ way to determine whether the record returned by `FindFile()`
was newly allocated or is owned by the `MCServerScript`.

`MCServerScript::FindFileIndex()` is used by the server debug
infrastructure while adding breakpoints.  Sometimes it is called to
add breakpoints in files that have not yet been loaded, which means it
needs to call `FindFile()` to create new records for unloaded files.

Coverity Scan detected that `FindFileIndex()` was always discarding
the return value of `FindFile()`, leading to memory leaks and broken
server debugging.

This patch works around the problem by making `FindFileIndex()` try to
detect if `FindFile()` allocated a new record and, if so, handing over
ownership to the `MCServerScript` instance.  However, this is fragile,
and it would be better to refactor `MCServerScript::FindFile()` (etc.)
to make this unnecessary.

Coverity-ID: 16332