build: Guess appropriate paths for Win SDKs in configure
On 64-bit Windows, the QuickTime and Microsoft Speech SDKs are
installed in an x86-specific 'Program Files' directory, usually
`C:\Program Files (x86)\`.  However, the default SDK paths in the
gyp files are defined relative to `C:\Program Files\`.

This isn't a problem on the Wine-based Windows builders, where
the SDK are installed in `C:\Program Files\` as expected, or when
configuring via `configure.bat`, which generates suitable arguments
for `config.py`.  However, on Windows Server 2016, the build was
looking in the wrong place.

This patch makes `config.py` apply the same logic as
`configure.bat` to look for the QuickTime and Microsoft Speech
SDKs.