build: Launch mspdbsrv (well) outside build tree
The first step in each buildbot run is to clean the working tree
from the last run.  On Windows, it is not possible to delete a
directory (or its parents) which is still the current working
directory for a process.

The `make.cmd` batch script launches the `mspdbsrv` service to
collect debug information from the Windows build.  The service
runs continuously in the background; it isn't shut down when the
build completes.

Since the `make.cmd` script was being run with its CWD in the build
tree, `mspdbsrv` was also remaining running in the build tree,
blocking the deletion of the build tree.

`mspdbsrv` can be run with any CWD, so this patch modifies
`make.cmd` to launch `mspdbsrv` in the `$ProgramFiles(x86)`
directory.