[Bug 18912] Lock the target for execution while handling a message
Previously it was possible to delete the target object if the
message was not handled or had not yet been handled by it. This
patch will ensure that whenever the target is changed it is
locked for execution. Any code that deletes the target or
its parent in a front script or deletes the target using something
like `send "delete me" to me in 0` will need to be updated.