build: Manage mspdbsrv instances during Win builds
Builds were stalling and timing out on Windows, and this appeared
to be due to `mspdbsrv` instances hanging around after the build
had been completed.

Further investigation revealed a host of issues related to
`mspdbsrv` process management.  This patch updates the Windows
buildbot build script to prevent as many of these issues as
possible.