[[ SocketLocalPort ]] Added the ability to specify the local host and port when opening a new socket.
Updated the parsing of open socket so that it now accepts the following form:
open socket [from [<local host>][:<local port>]] to <host>[:<port>]

Reworked the name to sockaddr parsing into:
- name to host and port
- host and port to sockaddr
Name to sockaddr now calls name to host and port and then host and port to
sockaddr.

The reworked name to sockaddr parsing is used to parse any local from details in
MCS_connect_socket: It checks to see if any local host or port has been passed.
If no local host has been passed, then we use, if set, the networkinterface
property. It then binds the socket to any found local host and port.