Improved check for fixed-width font
This commit fixes a bug where fixed width fonts were _never_ detected.
The length argument to `MCGContextMeasurePlatformText()` is the number
of _bytes_ in the text, rather than the number of characters.  The
fixed-width check was passing 1 as the text length, and `1 / 2` in
integer arithmetic is 0, so `MCGContextMeasurePlatformText()` was
always returning a width of 0 here.

The check for a fixed-width font in `MCFontCreateWithFontStruct()` can
be made considerably more efficient by unrolling the measurement of
the width of a space.  The body of the loop over characters is reduced
to a single function call and test.

On clang, this allows the compiler to unroll the loop completely in
release mode, and on GCC it eliminates a `-Wmaybe-uninitialized` error
for the `t_last_width` variable.