libgraphics: Add constructors to MCGPoint
In some places in libgraphics (e.g. `MCGPathIterate()`), `MCGPoint`
arrays are allocated on the stack.  Providing a default constructor
for `MCGPoint` ensures that the contents of these arrays gets
automatically initialized to default values.

Because there are also some places in the source code that construct
an `MCGPoint` with an initialiser list, we need a "normal" constructor
too.