[[ SysLibrary ]] Update libscript to use system library functions
This patch updates libscript to use the new system library
loading functions, removing the per-platform code which was
previously present.

Libraries are now all loaded via a callback function, which degrades
gracefully to using MCSLibraryCreateWithPath if none is set. This
means that lc-run can continue to function as it does now; whereas
all library loading in the engine is done using MCU_library_load.