engine: Add MCTristate type for things that need to be true AND false
Buttons are currently allowed to be in a "mixed" highlight state.
Previously the `Boolean` type has been overloaded for this, requiring
it to be an integer type rather than `bool`.

This patch introduces a new `MCTristate` type and modifies every place
in the engine that handles tri-state values to use it, simplifying
some code in the exec interface and providing additional compile-time
protection.