lc-compile: Ensure local variable is initialised at acquisition
GCC 6.3.1 reports that the `t_index` variable in
`CountSyntaxNodeMarks()` may be uninitialised in some codepaths.
Since it's initialised in every branch of the switch statement, and
the switch covers every possible `SyntaxNodeKind`, it shouldn't
normally be possible, but it's probably best that the function doesn't
return random numbers on encountering a dodgy syntax node.