libfoundation: Define byte_t to be unsigned char
We rely on 3 properties of the `byte_t` type:

1. The `sizeof` C++ operator must return object sizes in units of
   `byte_t`

2. Casting a pointer to a `byte_t*` allows examination of the object
   representation of the pointer's target object

3. A `byte_t` has 8 bits

Unless `byte_t` is either `char`, `signed char` or `unsigned char`,
properties 1 and 2 would not hold.  There is no requirement in C++
that `uint8_t` and `unsigned char` must be treated as the same type by
a C++ compiler.

This patch changes `byte_t` to directly alias `unsigned char`, and
adds a static assertion that there are 8 bits in a `char` on the
target platform.