[Bug 19390] Reposition native layer controls when resizing stacks
The LiveCode coordinate space has its origin at the top left of
the stack, but the Cocoa coordinate space's origin is at the
bottom left of the window.  This means that native layer controls
end up in the wrong position if they are not explicitly
repositioned during stack resizing.

In commit 8fe69c5c43f6, the stack root container view was modified
so as to reposition its contents.  Subsequently, in commit
8f1e4c1bd1cf groups were taught to keep the same coordinate system
as the outer window.  Unfortunately, groups were not configured to
propagate position changes to their subviews.  This caused a bug
where browsers and players would be positioned in the wrong place
during stack resize, but only if they were part of a group.

This patch teaches groups' container views to propagate position
changes to subviews.