Fix strict aliasing warnings in externalv0 string conversion
It's permitted to cast from a `char_t*` to a `char*`, but casting the
other way is undefined behaviour.  Casting between reference types is
always a strict aliasing error.

This patch fixes warnings by introducing a helper function,
`convert_from_string()`, which takes care of all the required pointer
juggling.