[Bug 19318] Fix loading custom props from pre-7.0 stack files
When loading legacy custom properties, they should be _added_ to the
current contents of the custom property set instead of replacing it.

This patch adds a new private `fetch_ensure()` method that returns the
internal array of the object property set, populating it with a new
empty array if necessary.