libfoundation: Add object hashes
There's a fairly common requirement to perform a byte hash on the C++
_object representation_ of a value.  Actual example from libfoundation:

    t_hash = MCHashBytesStream(t_hash, &self->alias.name,
                               sizeof(self->alias.name));

This patch adds two new template functions to libfoundation:
`MCHashObject<T>()` and `MCHashObjectStream<T...>()`.  These functions
take a reference to a value, and return the hash of its object
representation.  `MCHashObjectStream<T...>()` is variadic in order to
allow multiple values of different types to be ergonomically
integrated into a hash.

To ensure correctness, these templates are only enabled for
`TriviallyCopyable` types, i.e. types for which copying the bytes
occupied by the object in the storage is sufficient to produce another
object with the same value.

The example above is then replaceable by:

    t_hash = MCHashObjectStream(t_hash, self->alias.name);

For now, these are only for use within libfoundation due to their
dependence on `<type_traits>`, but they could be moved into the public
API at some point.