[[ MCplayers ]] Refactor MCplayers access MCtemplateplayer belongs to MCplayers, and gets destructed and re-constructed every time `reset the templateplayer` is called.
So the code that removes a player from MCplayers needs to happen
in the MCPlayer destructor. This means that when players have been
deleted and their proxies invalidated, it is possible for invalid
MCPlayerHandles to be part of the MCplayers list.

This patch refactors MCplayers access and, in particular, ensures
that no crash can occur when iterating through the MCplayers list.