build: Build only for specific target triples
Ensure that LiveCode only builds for specific target triples,
skipping the build for anything other than the specific triples
supported by the branch.

The target triple check is only for the configure and compile build
steps, because other build steps will either be automatically
skipped if configure skips, or because the steps are for creating
the installer and the target triple doesn't matter.