[[ Bug 19479 ]] Make modals close automatically on Cmd-.
This patch causes modal dialogs to be automatically closed if a
script abort occurs. This stops the possibility of the IDE getting
locked up due to a broken modal script.

If allowInterrupts is false, or the stack has cantAbort set to true
then the behavior is unaffected (i.e. Cmd-. will have no effect).