libfoundation: Add private in-place reverse template
Add a private template that provides in-place reverse for an array.
The algorithm works from the middle of the array, working outwards and
swapping elements between the front and the back.  It can only be used
for types that are suitable for use with `std::swap()`; either
trivially copyable, or with an implementation of `T::swap(T& other)`.