libfoundation: Add MCStringCopyReversed()
Strings cannot be reversed in-place in the general case.  There is no
upper limit on the number of codeunits in a grapheme, which means that
a temporary buffer of at the same length of the string is required for
general correctness.

This patch adds a "reversed copy" function which performs a deep copy
of the source string, and then constructs the reversed form in the
copy by re-reading from the source string via a grapheme break
iterator.

In practice, the "trivial" flag is not getting set for trivial
strings, so all strings that use UTF-16 internal representatino go
through the slow grapheme break iterator path.