lcb: Integer type improvements
This patch reworks the sized integer types ensuring that all types
which may be distinct are truly distinct. In doing so it unifies
the naming of all types (S prefix for signed integers, U for unsigned
integers), adds all the C integral types, and adds aliases from
Int* to SInt*.

A particular outcome of this patch is that foreign handlers which
bind to Java methods *must* use the J* types and not the C* types
as the C types are now considered distinct, and the Java binding
code expects the J types (which are just aliases to the machine
types since the JNI fixes that mapping).