[Bug 17097] libfoundation: Add MCMemoryClearSecure()
Add a new function, `MCMemoryClearSecure()`, that's equivalent to
`MCMemoryClear()` but with the guarantee that it will never be
optimised out.

On Windows, this uses the standard `SecureZeroMemory()` library
function.

On other platforms, this uses `MCMemoryClear()` (i.e. `memset()`), but
with an additional memory barrier that prevents the call to `memset()`
from being optimised out.  Typical assembly generated in `Release`
mode:

    push	rbx
    mov	rdx, rsi
    xor	esi, esi
    mov	rbx, rdi
    call	memset
    pop	rbx
    ret