[Bug 17540] Fix crash in LiveCode on Mac with -h
When `X_init()` fails, the Mac engine constructs an startup error
message using the contents of `MCresult`.  One of the reasons that
`X_init()` might fail is when it encounters `-h` in the command-
line arguments.  In this case, the Mac engine was crashing while
trying to access `MCresult`.

`X_open()` is responsible for initialising `MCresult`, and it is
called with the command-line arguments that are left over after the
engine has processed them.  When `-h` is encountered, `X_open()` is
never called.

This patch slightly improves the situation by ensuring that
`MCPlatformHandleApplicationStartup()` checks that `MCresult` is
initialised before attempting to retrieve its contents.  However,
it would be better to more clearly distinguish between printing the
usage message as requested and an actual error.