installer: Canonical extension module names are now lowercase
In commit c7b53c4a17d7, all extensions were updated to use a
lowercase name.  However, the package manifest was not updated with
the new names.  Since LiveCode installers are generated on a case-
sensitive filesystem, this caused some extensions to be omitted
from the installers, breaking the IDE.