[[ Bug 19488 ]] Separate engine path and code path
This patch makes a distinction between the engine path (MCcmd)
and the path to use to load shared libraries from (MCappcodepath).

These are the same on most platforms, but differ on emscripten
and android.

Currently all platforms have specific rules for computing MCcmd
and MCappcodepath:

- on Desktop, iOS and Server platforms MCcmd is the path to the
  executable and MCappcodepath is the folder containing it.

- on Android, MCcmd is the path to the APK, and MCappcodepath
  is the folder containing the engine library.

- on Emscripten, MCcmd is argv[0] and MCappcodepath is not used.

To allow the appcodepath to be configured before initializing the
engine, the declaration of the X_* lifecycle methods (which are
implemented separately for Desktop, Mobile, Server and Emscripten)
have been moved to globals.h, and X_init now takes a struct of
options, rather than an argument list.