[[ Emscripten ]] Fix emscripten engine
This patch makes sure the emscripten engine works with the new
appcode path changes and fixes an issue from the rework of library
loading which meant it failed to run.

In addition, it tweaks the libfoundation gyp file so the emscripten
engine can be built on Mac; and adjusts the testgen script for a
similar reason.