[Bug] Fix go card stack placement
In certain circumstances it was possible for go card to cause the
misplacement of a stack because the relative rect being used was
coming from an invisible script only stack. This is only ever
likely to be an issue when modal stacks are presented with no
other visible stack.

This patch ensures that the relative stack being used by go card
is the same as used by open stack.