[[ Bug 19593 ]] Ensure 'type' works correctly for accented chars
This patch ensure that the 'type' command uses the same logic as
normal keyboard entry. If a typed char is mappable to native, then
its keycode is the native code for the char; otherwise the keycode
is the unicode codepoint value with bit 24 set. In both cases, the
key string is the unicode character.