[[ Win64 ]] Sort out arch vars, and ensure prebuilts are fetched
This patch ensures that TARGET_ARCH is always x86_64 for 64-bit
X86 processors, and that target_arch (the gyp variable) is always
x64 for x86-64 on Win32. Additionally, it introduces a new gyp var
uniform_arch, which is either x86_64 or x86.

Finally, a 'fetch' phase has been added which fetches the prebuilts
for host and target. This ensures that they are present before
configuration (which eliminates missing file warnings from gyp) and
that Visual Studio can build projects in its own order (which can
cause failure to find prebuilt includes).