[[ Build ]] Ensure git_revision is correct
This patch ensures that the git_revision for commerical builds
is that of the private repo; whilst the git_revision for community
builds is that of the public repo.

This is achieved by getting config.py to set a build_edition variable
in gyp and then a condition in version.gypi which fetchs the git
revision from .. if commercial and . if community.