[[ Build ]] Ensure git_revision is correct
This patch computes the git_revision of the build at config time
by setting the git_revision git variable in the command line args
passed to gyp from config.py.

The git revision for a community build is that of the livecode
repository; for a commercial build it is that of the livecode-private
repository.