[[ Bug 19584 ]] Add implements of std::move and std::swap
This patch adds implementations of std::move and std::swap but as
std_::move and std_::swap to object.h. These are to ensure that the
code compiles on develop-8.1 which uses older toolchains (some of which
don't support C++11). In particular, the Mac build will fail without
these definitions.

Note: When merging this commit up, the std_ namespace definitions can
be removed, and uses of std_ replaced with std.