[Bug 18150] Add format string param to date function
This patch adds an additional optional parameter to the
date function allowing a format string to be specified.

Currently in order to format the current datetime in
a format that is not natively supported the user needs
to get the date in one of the supported formats,
convert to dateItems and then format in the required format.