[[ Android ]] Tweak android build readme
This patch corrects the command that should be used to build
an android standalone toolchain to match how android is built
on vulcan.

In particular, it is important that no --stl option is specified
otherwise the engine ends up with a dependence on libc++_shared.so
rather than linking to it statically.