[[ Bug 19635 ]] Fix native layer placement on Mac
This patch fixes an issue where if a native layer is in a group
then it would not update its y location correctly.

Previously, there was code in our NSView subclasses to deal
with the vertical height of the viewport changing (Cocoa's coord
system is bottom-left relative).

This code has been removed, and has been replaced by a call to
doSetGeometry() in doSetViewportGeometry() in the Mac native view
class.

Additionally, a missing call to MCControl::viewportchanged has been
added to MCGroup::viewportchanged.

There is a revbrowser-specific 'hack' in the MCWindowContainerView
class still - which ensures a revbrowser layer will remain at the
correct y-location when the height of a window changes.