[[ Bug 19587 ]] Relax object execution lock
This patch relaxes the object execution lock as far as it can.

Locks are only taken on an object if it is executing a handler,
or if one of its behavior's is executing a handler.

In all other cases (e.g. handling a message in an owning object
after the message passes), deleting the target will work, but
the target will be unset afterwards.

If the target is deleted at any point in the message path,
the current message will not pass - regardless of whether
pass was invoked or not.

This is safe as any attempt to use 'the target' after it has
been deleted will throw an error as there is no object to
resolve anymore.