[Bug 19541] Improve clipboard owner check on win
This patch modifies the windows clipboard ownership
check to return true for a cleared or new raw clipboard
item so private clipboard data set on an otherwise
cleared clipboard is not inadvertently cleared.

A NULL m_object of the raw clipboard item is owned in the
sense that it should only exist if the clipboard has been
set.