[[ Bug 11323 ]] Add difference and symmetric difference
This patch implements 'difference' and 'symmetric difference'
commands for arrays.

The difference command removes all keys from the destination
which are present in the source, and leaves all others alone.

The symmetric difference command removes all keys from the
destination which are present in the source, and adds all keys
from the source which are not present in the destination.