[Bug 19541] Fix clipboard `IsEmpty()`
A previous patch added `CreateNewItem()` to the raw
clipboard `Clear()` method. This is incorrect as
a cleared clipboard is distinct from a clipboard
containing a single item with a `NULL` `m_object`.

The error caused a regression in the drag behavior
that uses the clipboard `IsEmpty()` method to
determine if a drag has started.