Update print.lcdoc to address issue in bug #19570
I have changed the references to topLeft to leftTop as identified.   The leftTop parameter did not have an explanation.  The rightBottom parameter had the explanation for both parameters but was not clear.  I edited the rightBottom parameter explanation to focus on that parameter specifically and added a leftTop description to do the same thing.  Please check to see that the descriptions of the parameters are correct.