[Bugfix 19570] Tweak usage of bottomLeft/topRight in print docs (#5486)
* Create file bugfix-19803.md

bugfix-19803.md added through web interface

* Revert "Create file bugfix-19803.md"

* Update print.lcdoc to address issue in bug #19570

I have changed the references to topLeft to leftTop as identified.   The leftTop parameter did not have an explanation.  The rightBottom parameter had the explanation for both parameters but was not clear.  I edited the rightBottom parameter explanation to focus on that parameter specifically and added a leftTop description to do the same thing.  Please check to see that the descriptions of the parameters are correct.

* Update print.lcdoc to address bug #19570

There were two entries for the leftTop parameter.  One was blank and one was the last parameter before the description.  I have already added text to the first entry.  I have deleted the second entry.

* Fixing minor mistake an formatting

Ensuring the wording and formatting of the leftTop and rightBottom is consistent.

* Update print.lcdoc

Changed leftTop to topLeft and rightBottom to bottomRight in response to comments from reviewers.  Added a sentence clarifying the corrordinate order.  I copied that sentence from the mobileAdCreate doc entry.

* added bugfix release note

added release note bugfix-19570.md