[[ Bug 18293 ]] Clear MCmenuobjectptr
The patch for bug 18293 leaves the button as the menu object
but the popup stack has been deleted. This means the button
can not be deleted itself. This patch checks if the button
is the menu object when a menu is being deleted and sets
`MCmenuobjectptr` to `nil`.

(cherry picked from commit 167281c65d36f2cbd3901a6e965ef230fa5f1482)