[[ Toolchain ]] Use ordinals for binding to builtin handlers
This patch modifies lc-compile's outputc mode so that there is
no need to use dlsym to lookup any builtin foreign handlers which
are referenced by any of the generated modules.

This works by generating a list of all builtin handlers as
the collection of modules are compiled. All 'builtin' foreign
handler definitions have their binding string changed to a non-negative
integer which indexes into a C array of function pointers which
is generated before the module loader classes in the generated
C++ file.